+++ js/app/components/About.react.js
@@ -1,0 +1,12 @@
+var
+  React = require("react"),
+  About = React.createClass({
+    render: function() {
+      return (
+        <h2>About page</h2>
+      );
+    }
+  });
+
+
+module.exports = About;