+++ js/app/components/KanbanerApp.react.js
@@ -1,19 +1,56 @@
-var React = require('react');
+var
+  React = require( "react" ),
+  Router = require( "react-router" ),
+  RouteHandler = Router.RouteHandler,
 
-var Navigation = require('./Navigation.react');
-var MainSection = require("./MainSection.react");
-var Footer = require('./Footer.react');
+  UserStore = require( "../stores/UserStore" ),
 
-var KanbanerApp = React.createClass({
-  render: function() {
-    return (
-      <div>
-        <Navigation/>
-        <MainSection/>
-        <Footer/>
-      </div>
-    );
-  }
-});
+  Navigation = require( "./Navigation.react" ),
+  MainSection = require( "./MainSection.react" ),
+  Footer = require( "./Footer.react" );
+
+  getState = function() {
+    return {
+      isUserLoggedIn: UserStore.isUserLoggedIn()
+    };
+  };
+
+  KanbanerApp = React.createClass(
+    {
+      mixins: [Router.Navigation],
+
+      getInitialState: function() {
+        return getState();
+      },
+
+      componentDidMount: function() {
+        UserStore.addUserLoggedOutListener( this.onUserLoggedOut );
+      },
+
+      compoenentWillUnmount: function() {
+        UserStore.removeUserLoggedOutListener( this.onUserLoggedOut );
+      },
+
+      render: function() {
+        return (
+          <div>
+            <Navigation/>
+            <RouteHandler />
+            <Footer/>
+          </div>
+        );
+      },
+
+      /**
+       * Handler for USER_LOGGED_OUT events
+       */
+      onUserLoggedOut: function() {
+        this.setState( getState() );
+        if( !this.state.isUserLoggedIn ) {
+          this.transitionTo( "/login" );
+        }
+      }
+    }
+  );
 
 module.exports = KanbanerApp;